add all authentification process with password

+++ src/modules/auth/repository.js
@@ -11,7 +11,7 @@ function authRepository(database) {
 
   function getCredentials(customerId) {
     return database.query(`
-      SELECT customer_id::integer AS "customerId", password
+      SELECT customer_id AS "customerId", password
       FROM credentials
       WHERE customer_id = $1
       LIMIT 1
